Refactored responsiblities between base and derived Framework types

+++ Engine/src/Time.h
@@ -2,5 +2,8 @@
 
 namespace ZeroEngine {
 
-    typedef long Time;
+    // @@TODO: Make class so I can have methods and statics
+    // associated with Time
+    typedef double Time;
+
 }
